refactor: View에서 인터페이스 도입
