+++ src/main/java/racingcar/Application.java
@@ -1,13 +1,13 @@
 package racingcar;
 
 import racingcar.controller.RacingCarController;
-import racingcar.view.InputView;
-import racingcar.view.OutputView;
+import racingcar.view.InputViewImpl;
+import racingcar.view.OutputViewImpl;
 
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        RacingCarController racingCarController = new RacingCarController(new InputView(), new OutputView());
+        RacingCarController racingCarController = new RacingCarController(new InputViewImpl(), new OutputViewImpl());
         racingCarController.run();
     }
 }
