+++ src/main/java/racingcar/view/InputView.java
@@ -1,20 +1,8 @@
 package racingcar.view;
 
-import camp.nextstep.edu.missionutils.Console;
+public interface InputView {
 
-public class InputView {
+    String readCars();
 
-    public String readCars() {
-        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-        return readLine();
-    }
-
-    public String readRound() {
-        System.out.println("시도할 회수는 몇회인가요?");
-        return readLine();
-    }
-
-    protected String readLine() {
-        return Console.readLine();
-    }
+    String readRound();
 }
