+++ src/main/java/racingcar/view/InputViewImpl.java
@@ -1,0 +1,20 @@
+package racingcar.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class InputViewImpl implements InputView {
+
+    public String readCars() {
+        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
+        return readLine();
+    }
+
+    public String readRound() {
+        System.out.println("시도할 회수는 몇회인가요?");
+        return readLine();
+    }
+
+    protected String readLine() {
+        return Console.readLine();
+    }
+}
