+++ src/main/java/racingcar/view/OutputView.java
@@ -1,18 +1,9 @@
 package racingcar.view;
 
-import racingcar.model.Racer;
+public interface OutputView {
+    void printExecution();
 
-public class OutputView {
+    void printResult(String result);
 
-    public void printExecution() {
-        System.out.println("실행 결과");
-    }
-
-    public void printResult(Racer racer) {
-        System.out.println(racer.toString());
-    }
-
-    public void printFinalWinner(String winner) {
-        System.out.println("최종 우승자 : " + winner);
-    }
+    void printFinalWinner(String winner);
 }
